# dotnet-video-center/ClientApp/src/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Media And Tag API",
    "description": "Manage your local media files.",
    "contact": {
      "name": "mushroom-cn",
      "url": "https://github.com/mushroom-cn/dotnet-video-center"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/mushroom-cn/dotnet-video-center/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/v1/Media/{mediaId}/tag/{tagId}": {
      "delete": {
        "tags": [
          "Media"
        ],
        "summary": "Unbind tag with media.",
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "description": "Media id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "description": "Tag id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Media"
        ],
        "summary": "Bind tag with media.",
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "description": "Media id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "description": "Tag id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/v1/Setting/{mediaId}/tag/{tagId}": {
      "delete": {
        "tags": [
          "Setting"
        ],
        "summary": "Unbind tag with media.",
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "description": "Media id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "description": "Tag id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Setting"
        ],
        "summary": "Bind tag with media.",
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "description": "Media id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "description": "Tag id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/Tag/id/{id}": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Get Tag info by tag id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tag id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Tag": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Get Tag info collection by a group of tag id",
        "requestBody": {
          "description": "A group of tag id",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tag"
        ],
        "summary": "Modify Tag",
        "requestBody": {
          "description": "A group of tag",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tag"
        ],
        "summary": "Create Tag",
        "requestBody": {
          "description": "A group of tag",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "Delete Tag by a group od id",
        "requestBody": {
          "description": "A group of tag id",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/Tag/name/{name}": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Get Tag info collection by tag name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Tag name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Media": {
        "required": [
          "iconUrl",
          "mediaType",
          "name",
          "path"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 250,
            "type": "string"
          },
          "path": {
            "maxLength": 300,
            "type": "string"
          },
          "mediaType": {
            "maxLength": 30,
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifyTime": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "mediaTags": {
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaTag"
            },
            "nullable": true
          },
          "iconUrl": {
            "type": "string"
          },
          "url": {
            "maxLength": 2083,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MediaTag": {
        "required": [
          "mediaId",
          "tagId"
        ],
        "type": "object",
        "properties": {
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "tagId": {
            "type": "string",
            "format": "uuid"
          },
          "tag": {
            "$ref": "#/components/schemas/Tag"
          },
          "mediaId": {
            "type": "string",
            "format": "uuid"
          },
          "media": {
            "$ref": "#/components/schemas/Media"
          }
        },
        "additionalProperties": false
      },
...
      "Tag": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 250,
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifyTime": {
            "type": "string",
            "format": "date-time"
          },
          "mediaTags": {
            "maxItems": 999,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaTag"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}